indian_market_news/requirements.txt
requests==2.26.0
python-dotenv==0.19.0
schedule==1.2.1
gnews==0.3.6